Sixth Commit:
Implementing the rest of log data functionality for Contract (create, delete, patch), functionality also creating Updating ContractEvent for Contracts affected from updates and deletes in customers, resellers or subscriptions.

+++ src/main/java/ivar/hogblom/crmbackend/dto/SubscriptionForContractComponentsDto.java
@@ -11,5 +11,5 @@ public record SubscriptionForContractComponentsDto(
         Integer contractLength,
         Integer renewalPeriod,
         Boolean active,
-        Integer pricePerMonth
+        Double pricePerMonth
 ) {}
